Switch from HTTP API to SOAP API.
Rewritten gem a bit in internals, so now it uses Bing Translator SOAP API instead of HTTP API. Interface are left intact.

Reason is that's not practically possible to send maximum allowed 10000 symbols to Bing Translator, as for HTTP API web servers are refuses to process the request with 414 "Request is too long" error. Practical limit is about 2000 bytes, and for non-latin scripts this limit become very small if count in a symbols due to URL encoding. For example, 1 cyrillic symbol becomes 2 bytes in UTF-8 and 6 bytes when URL encoded (Cyrillic letter Я becomes %D0%AF, and CJK hieroglyph 私 becomes %E7%A7%81 - 9 bytes!).

As SOAP API uses XML-formatted requests that are sent to the server with POST method - this problem is gone.

I've written tests for long texts and html input/output.